+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Version 1.1.2 (2021-07-05)
 1. Updated ``sc.heading()`` to handle arguments the same way as ``print()``, e.g. ``sc.heading([1,2,3], 'is a list')``.
 1. Updated ``sc.commaticks()`` to use locale-aware formatting. 
 1. Removed the ``fig`` argument from ``sc.commaticks()`` and ``sc.SIticks()``; now, the first argument can be an ``Axes`` object, a ``Figure`` object, or a list of axes.
+1. Updated ``sc.suggest()`` to modify the threshold to be based on the length of the input word.
 
 
 Version 1.1.1 (2021-03-17)